+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Changelog
 Some notable infrastructure contributions below. See [commits](commits/master) for Slack API updates.
 
 * Your contribution here.
+* Added events - [@dblock](https://github.com/dblock).
 * Organized methods into groups, added scraped group descriptions - [@dblock](https://github.com/dblock).
 * Replaced manually edited docs with a rake task `api:methods:update` that parses API at https://api.slack.com - [@rusikf](https://github.com/rusikf).
 * Initial fork, updated December 22, 2014 from https://github.com/slackhq/slack-api-docs - [@dblock](https://github.com/dblock).